Update README.md
Saw this on reddit. Anyway, the numbering of the ordered list in the section *Several ways to use the class:* seems off. Every example is numbered *1*. If this is unintentional, this PR resolves the problem. Also, as an FYI, you're missing the commit(s) that add the actual code. Only the initial commit is there, which just has `ReplaceMe.m`. I can open an issue for that if you like. Thanks for sharing the library.